Extract Windows1256 to its own file
